Workaround inverted MF drive
Previous builds implemented manual setup mode by toggling LV on, driving
focus & toggling LV off, so that the camera viewfinder could be used to
confirm focus. Occasionally the direction of focus would get inverted
when reversing directions. This seemed to be a more frequent problem
with non-AFS lenses.

In this checkin, the manual mode requires the user to enable LV
explicitly, drive focus & disable LV to use the viewfinder.

LV state is toggled with the select button in the manual handler class
when the active field is "Test". When the active field is not "Test",
the select button returns to the main menu (as in previous builds).
TODO: remove some serial prints.